review fixes: move wallet_file_path outside ctx impl
